Moved blog at /index to /blog
