+++ components/landing-layout.js
@@ -1,0 +1,13 @@
+import Head from 'next/head';
+import styles from './layout.module.css';
+import utilStyles from '../styles/utils.module.css';
+import Link from 'next/link';
+import MenuIcon from './menu-icon';
+
+export default function LandingLayout({ children }) {
+    return (
+        <div>
+
+        </div>
+    );
+}
